/********** Item detected when clicked and moveable *************/ ------------------------------------------------------------------
OnClick, the listbox items retrieve their names from an array based
upon the index that is returned by the controlEvent. Te item is then
removed from the original list and then added to a new list (which
represents the canvas). A "copy" of the item is also added to the
pointer as feedback to drop onto the canvas.