/*       Item draggable to any location on the applet window        */  ------------------------------------------------------------------
OnClick, the listbox items retrieve their names from an array based
upon the index that is returned by the controlEvent. The item is then
removed from the original list and then added to a new list (which
represents the canvas). The item can now be dragged to any location
on the applet window, but only ON THE FIRST DRAG. Once released,
the item's position is SET and it cannot be MOVED unless the user uses
Alt+drag.

There is currently NO feedback DURING the drag.
(Nothing follows the cursor)

Next step is to make it so that the item can ONLY be dragged onto
the canvas area. If the item is dragged OUTSIDE of the canvas area
the color of the item SHOULD NOT CHANGE.

/******** ControlP5 Library updated to 0.6.0 release **************/