/*     Item draggable to dropCanvas target, now with Cursor     */ ------------------------------------------------------------------
OnClick, the listbox items retrieve their names from an array based
upon the index that is returned by the controlEvent. The item is then
removed from the original list and then added to a new list (which
represents the canvas). The item can now be dragged to any location
on the applet window, but only ON THE FIRST DRAG. Once released,
the item's position is SET and it cannot be MOVED unless the user uses
Alt+drag.

Mini Update: the cursor now works. Item can ONLY be dragged onto
the dropCanvas area.

Bugs: If the user performs a drag before selecting an item, there will
be a NULL POINTER response. There should be a check for having clicked
on an item, but not released.

Next step: fix bugs, and if the item is dragged OUTSIDE of the canvas area
the color of the item SHOULD NOT CHANGE.