/*    Item draggable to dropCanvas target: BUGFIXES COMPLETE    */ ------------------------------------------------------------------
OnClick, the listbox items retrieve their names from an array based
upon the index that is returned by the controlEvent. The item is then
removed from the original list and then added to a new list (which
represents the canvas). The item can now be dragged to any location
on the applet window, but only ON THE FIRST DRAG. Once released,
the item's position is SET and it cannot be MOVED unless the user uses
Alt+drag.

/****************************   UPDATE   ****************************/
Cursor no longer appears when user is collapsing/expanding the list.
There is no longer a NULL POINTER error when user performs a drag
without an item. Color of the item no longer changes when the item
is dragged outside of the dropCanvas.

/***********************II*   NEXT  STEP  **************************/
When users click on an item that is ALREADY in dropCanvs, the item
should be removed from dropCanvas, and the item color returned to
it's original state.