/* Item is removed from Canvas on second click + BUGFIX */ ------------------------------------------------------------------
OnClick, the listbox items retrieve their names from an array based
upon the index that is returned by the controlEvent. The item is then
removed from the original list and then added to a new list (which
represents the canvas). The item can now be dragged to any location
on the applet window, but only ON THE FIRST DRAG. Once released,
the item's position is SET and it cannot be MOVED unless the user uses
Alt+drag.

/*************************   UPDATE   ****************************/
  When users click on an item that is ALREADY in dropCanvs, the item
  should be removed from dropCanvas, and the item color returned to
  it's original state.


/***********************   NEXT  STEP  **************************/
  Make it so that the dropCanvas recognizes the (currently very
  simple objects) and stores those into an ArrayList.