/* Item is removed from Canvas on second click + BUGFIX */ …  ------------------------------------------------------------------
OnClick, the listbox items retrieve their names from an array based
upon the index that is returned by the controlEvent. The item is then
removed from the original list and then added to a new list (which
represents the canvas). The item can now be dragged to any location
on the applet window, but only ON THE FIRST DRAG. Once released,
the item's position is SET and it cannot be MOVED unless the user uses
Alt+drag.

/*************************   UPDATE   ****************************/
  Code has been sorted and cleaned into two classes. A new class
  has been added called FileObject. FileObject is actually how we
  create each <file></file> for the xml. It contains all of the
  attributes that the file should have, including filename, trigger,
  size (w&h), position(x&y), scene, and transition. These fileObjects
  will be stored into a global arrayList, titled fileObjectArray.


/***********************   NEXT  STEP  **************************/
  Write a update() function for fileobject, and make it so that
  the properties of the fileObject can be printed. Integrate code
  with Ashton's XML reading!